Refactored Metric
