+++ monitor/src/test/java/bisq/monitor/MonitorInfrastructureTests.java
@@ -26,8 +26,6 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
 
-import bisq.monitor.metric.Metric;
-
 public class MonitorInfrastructureTests {
 
     /**
