+++ monitor/src/test/java/bisq/monitor/TorHiddenServiceStartupTimeTests.java
@@ -30,7 +30,6 @@ import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
 
-import bisq.monitor.metric.Metric;
 import bisq.monitor.metric.TorHiddenServiceStartupTime;
 
 public class TorHiddenServiceStartupTimeTests {
